calims2-webapp/test/unit/java/gov/nih/nci/calims2/ui/administration/customerservice/invoice/InvoiceFormTest.java
/*L
 * Copyright Moxie Informatics.
 *
 * Distributed under the OSI-approved BSD 3-Clause License.
 * See http://ncip.github.com/calims/LICENSE.txt for details.
 */

/**
 *
 */
package gov.nih.nci.calims2.ui.administration.customerservice.invoice;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

import org.junit.Test;

import gov.nih.nci.calims2.domain.administration.customerservice.CustomerRequest;
import gov.nih.nci.calims2.domain.administration.customerservice.Invoice;
import gov.nih.nci.calims2.domain.administration.customerservice.enumeration.InvoiceStatus;
import gov.nih.nci.calims2.domain.common.Type;
import gov.nih.nci.calims2.ui.generic.crud.CRUDFormAssert;
import gov.nih.nci.calims2.ui.generic.crud.CRUDFormTestHelper;

/**
 * @author katariap
 *
 */
public class InvoiceFormTest {

  private static final Long[] IDS = new Long[] {1L, 2L};

  /**
   * Test the controller getSubmittedEntity method.
   */
  @Test
  public void testGetSubmittedEntity() {
    InvoiceForm form = new InvoiceForm();
    form.setCustomerRequestIds(IDS);
    CRUDFormTestHelper.setNotes(form);
    form.setStatus(InvoiceStatus.DEFAULT.name());
    form.setTypeId(1L);
    Invoice entity = form.getSubmittedEntity();
    assertNotNull("No submitted Entity", entity);
    CRUDFormAssert.assertEntities(CustomerRequest.class, entity.getCustomerRequestCollection(), IDS);
    CRUDFormTestHelper.assertNotes(entity);
    assertEquals("Wrong status in the entity", InvoiceStatus.DEFAULT, entity.getStatus());
    CRUDFormAssert.assertEntity(Type.class, entity.getType(), 1L);
  }
}
